Move Select component stories into own js file
We'll get more stories in future. Therefore seperate them by component.
