+++ ng/src/web/stories/config.js
@@ -20,97 +20,14 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-import React from 'react';
+import {configure} from '@storybook/react';
 
-import glamorous from 'glamorous';
-
-import {configure, storiesOf} from '@storybook/react';
-import {action} from '@storybook/addon-actions';
-
-import Select from '../components/form/select.js';
-
-const Sizer = glamorous.div({
-  width: '300px',
-});
-
-class ControlledSingleSelect extends React.Component {
-
-  constructor(...args) {
-    super(...args);
-
-    this.state = {value: 'ipsum'};
-
-    this.handleChange = this.handleChange.bind(this);
-    this.action = action('state change');
-  }
-
-  handleChange(...args) {
-    this.action(...args);
-    const [value] = args;
-    this.setState({value});
-  }
-
-  render() {
-    const {value} = this.state;
-    return (
-      <Sizer>
-        <Select
-          value={value}
-          onChange={this.handleChange}
-        >
-          <option value="foo">Foo</option>
-          <option value="bar">Bar</option>
-          <option value="lore">Lore</option>
-          <option value="ipsum">Ipsum</option>
-        </Select>
-      </Sizer>
-    );
-  }
-}
+/* eslint-disable global-require, no-console */
 
 function loadStories() {
-  console.clear(); // eslint-disable-line no-console
+  console.clear();
 
-  storiesOf('Select', module)
-    .add('with options', () => (
-      <Sizer>
-        <Select
-          onChange={action('select value change')}
-        >
-          <option>Foo</option>
-          <option>Bar</option>
-        </Select>
-      </Sizer>
-    ))
-    .add('with options (value and label)', () => (
-      <Sizer>
-        <Select
-          onChange={action('select value change')}
-        >
-          <option value="foo">Foo</option>
-          <option value="bar">Bar</option>
-          <option value="lore">Lore</option>
-          <option value="ipsum">Ipsum</option>
-        </Select>
-      </Sizer>
-    ))
-    .add('with controlled input', () => (
-      <ControlledSingleSelect/>
-    ))
-    .add('disabled controlled input', () => (
-      <Sizer>
-        <Select
-          value="bar"
-          disabled={true}
-          onChange={action('select value change')}
-        >
-          <option value="foo">Foo</option>
-          <option value="bar">Bar</option>
-          <option value="lore">Lore</option>
-          <option value="ipsum">Ipsum</option>
-        </Select>
-      </Sizer>
-    ));
+  require('./select');
 }
 
 configure(loadStories, module);
